Refactor: Reusable modal for exports based on category id
